# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mystics Inventory API — multi-tenant inventory SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/current:
    get:
      operationId: getCurrentOrganization
      tags: [organizations]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
    patch:
      operationId: updateCurrentOrganization
      tags: [organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrganizationBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
...
  /organizations/current/barcode-settings:
    patch:
      operationId: updateOrganizationBarcodeSettings
      tags: [organizations]
      summary: Update per-org barcode auto-generation prefix and format
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBarcodeSettingsBody" }
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Organization" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /pos/items/lookup:
    get:
      operationId: lookupPosItems
      tags: [pos]
      parameters:
        - { in: query, name: q, required: false, schema: { type: string } }
        - { in: query, name: warehouseId, required: false, schema: { type: integer } }
        - { in: query, name: limit, required: false, schema: { type: integer } }
        - { in: query, name: saleChannel, required: false, schema: { type: string }, description: "Resolve the warehouse from this channel's default before falling back to the org default." }
        - { in: query, name: bags, required: false, schema: { type: string }, description: "Set to '1' to return only bag items (isBag=true); q is not required when bags=1." }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PosLookupResult" }
  /pos/checkout:
    post:
      operationId: posCheckout
      tags: [pos]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PosCheckoutBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PosCheckoutResult" }
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
      required: [error]
...
    Organization:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        currency: { type: string }
        timezone: { type: string }
        gstNumber: { type: ["string", "null"] }
        addressLine1: { type: ["string", "null"] }
        addressLine2: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        postalCode: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        invoiceFooter: { type: ["string", "null"] }
        plan: { type: string }
        subscriptionStatus: { type: string }
        currentPeriodEnd: { type: ["string", "null"] }
        onboardingCompletedAt: { type: ["string", "null"] }
        barcodePrefix:
          type: ["string", "null"]
          description: Per-org prefix prepended to every auto-generated barcode. When null/empty, the generator falls back to a slug-derived default.
        barcodeFormat:
          type: string
          description: Active barcode format. Currently locked to "code128"; the column exists so EAN-13 / UPC-A can be added later without a migration.
        createdAt: { type: string }
      required: [id, name, slug, currency, timezone, plan, subscriptionStatus, createdAt, gstNumber, addressLine1, addressLine2, city, state, postalCode, country, logoUrl, invoiceFooter, currentPeriodEnd, onboardingCompletedAt, barcodePrefix, barcodeFormat]
    UpdateBarcodeSettingsBody:
      type: object
      properties:
        barcodePrefix:
          type: ["string", "null"]
          description: 1-8 alphanumeric characters. Pass null or empty string to fall back to the slug-derived default.
        barcodeFormat:
          type: string
          enum: [code128]
          description: Currently locked to code128.
...
    UpdateOrganizationBody:
      type: object
      properties:
        name: { type: string }
        currency: { type: string }
        timezone: { type: string }
        gstNumber: { type: ["string", "null"] }
        addressLine1: { type: ["string", "null"] }
        addressLine2: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        postalCode: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        invoiceFooter: { type: ["string", "null"] }
...
    PosLookupItem:
      type: object
      properties:
        id: { type: integer }
        sku: { type: string }
        name: { type: string }
        barcode: { type: ["string", "null"] }
        salePrice: { type: string }
        taxRate: { type: string }
        unit: { type: string }
        imageUrl: { type: ["string", "null"] }
        isBundle: { type: boolean }
        isBag: { type: boolean }
        trackBatches: { type: boolean }
        onHand: { type: number }
        maxDiscountPercent: { type: ["number", "null"] }
      required:
        [
          id,
          sku,
          name,
          barcode,
          salePrice,
          taxRate,
          unit,
          imageUrl,
          isBundle,
          isBag,
          trackBatches,
          onHand,
          maxDiscountPercent,
        ]
    PosLookupResult:
      type: object
      properties:
        warehouseId: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/PosLookupItem" }
      required: [warehouseId, items]
    PosCheckoutLine:
      type: object
      properties:
        itemId: { type: integer }
        quantity: { type: number }
        unitPrice: { type: number }
        taxRate: { type: number }
        discountPercent:
          type: number
          minimum: 0
          maximum: 100
          description: Per-line discount percent (0-100). Applied before tax. If set together with discountAmount, percent wins.
        discountAmount:
          type: number
          minimum: 0
          description: Per-line flat discount in rupees. Ignored when discountPercent is set.
        description: { type: ["string", "null"] }
      required: [itemId, quantity]
    PosCheckoutPayment:
      type: object
      properties:
        mode:
          type: string
          enum: [cash, card, upi, bank, other]
        amount: { type: number }
        referenceNumber: { type: ["string", "null"] }
        bankAccountLabel: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
      required: [mode, amount]
    PosCheckoutBody:
      type: object
      properties:
        lines:
          type: array
          items: { $ref: "#/components/schemas/PosCheckoutLine" }
        customerId: { type: ["integer", "null"] }
        warehouseId: { type: ["integer", "null"] }
        payment: { $ref: "#/components/schemas/PosCheckoutPayment" }
        notes: { type: ["string", "null"] }
        customerName: { type: string, minLength: 1 }
        customerPhone: { type: string, minLength: 1 }
        saleChannel:
          type: ["string", "null"]
          enum: [walkin, website, store, whatsapp, phone, instagram, other, null]
          description: Mode of sale captured at the POS — used for channel-level reporting.
      required: [lines, payment, customerName, customerPhone]
    PosCheckoutResult:
      type: object
      properties:
        salesOrderId: { type: integer }
        orderNumber: { type: string }
        customerId: { type: integer }
        warehouseId: { type: integer }
        customerPaymentId: { type: integer }
        receiptUrl: { type: string }
        subtotal: { type: string }
        taxTotal: { type: string }
        total: { type: string }
      required:
        [
          salesOrderId,
          orderNumber,
          customerId,
          warehouseId,
          customerPaymentId,
          receiptUrl,
          subtotal,
          taxTotal,
          total,
        ]