# lib/api-zod/src/generated/types/posLookupItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mystics Inventory API — multi-tenant inventory SaaS
 * OpenAPI spec version: 0.1.0
 */

export interface PosLookupItem {
  id: number;
  sku: string;
  name: string;
  /** @nullable */
  barcode: string | null;
  salePrice: string;
  taxRate: string;
  unit: string;
  /** @nullable */
  imageUrl: string | null;
  isBundle: boolean;
  isBag: boolean;
  trackBatches: boolean;
  onHand: number;
  /** @nullable */
  maxDiscountPercent: number | null;
}